+++ packages/mock-server/src/create-mock-server.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import { createMockServer } from './createMockServer.ts'
+import { createMockServer } from './create-mock-server.ts'
 
 describe('createMockServer', () => {
   it('GET /foobar -> example JSON', async () => {
