+++ packages/mock-server/src/index.ts
@@ -1,1 +1,1 @@
-export { createMockServer } from './createMockServer.ts'
+export { createMockServer } from './create-mock-server.ts'
