+++ packages/mock-server/src/routes/respond-with-authorize-page.ts
