+++ packages/mock-server/src/routes/respond-with-open-api-document.ts
