+++ packages/mock-server/src/routes/respond-with-token.ts
