+++ packages/mock-server/src/utils/find-preferred-response-key.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import { findPreferredResponseKey } from './findPreferredResponseKey.ts'
+import { findPreferredResponseKey } from './find-preferred-response-key.ts'
 
 describe('findPreferredResponseKey', () => {
   it('returns default over 200', () => {
