+++ packages/mock-server/src/utils/get-open-auth-token-urls.ts
