+++ packages/mock-server/src/utils/handle-authentication.ts
