+++ packages/mock-server/src/utils/is-authentication-required.ts
