+++ packages/mock-server/src/utils/log-authentication-instructions.ts
@@ -1,6 +1,6 @@
 import type { OpenAPIV3_1 } from '@scalar/openapi-types'
 
-import { getPathFromUrl } from './getOpenAuthTokenUrls.ts'
+import { getPathFromUrl } from './get-open-auth-token-urls.ts'
 
 /**
  * Log authentication instructions for different security schemes
