+++ packages/void-server/src/utils/create-json-response.ts
