+++ packages/void-server/src/utils/create-stream-response.ts
