+++ packages/void-server/src/utils/create-xml-response.ts
